remove pdiag_basic.cpp
