+++ source/src_pdiag/pdiag_double.cpp
@@ -699,3 +699,608 @@ void Pdiag_Double::diago_complex_begin(
 }
 
 
+
+#ifdef __MPI
+#include "../src_pw/occupy.h"
+void Pdiag_Double::gath_eig_complex(MPI_Comm comm,int n,std::complex<double> **cc,std::complex<double> *Z, const int &ik)
+{
+    ModuleBase::TITLE("Pdiag_Double","gath_eig_complex");
+    time_t time_start = time(NULL);
+    //GlobalV::ofs_running << " Start gath_eig_complex Time : " << ctime(&time_start);
+    const Parallel_Orbitals* pv = this->ParaV;
+
+    int i,j,k;
+    int nprocs,myid;
+    MPI_Status status;
+    MPI_Comm_size(comm,&nprocs);
+    MPI_Comm_rank(comm,&myid);
+
+    std::complex<double> **ctot;
+
+	// mohan add 2010-07-03
+	// the occupied bands are useless
+	// for calculating charge density.
+	if(GlobalV::DRANK> pv->lastband_in_proc)
+	{
+		delete[] Z;
+	}
+
+	// first we need to collect all
+	// the occupied bands.
+	// GlobalV::NBANDS * GlobalV::NLOCAL	
+	if(GlobalV::DRANK==0)
+	{
+		ctot = new std::complex<double>*[GlobalV::NBANDS];
+    	for (int i=0; i<GlobalV::NBANDS; i++)
+    	{
+        	ctot[i] = new std::complex<double>[GlobalV::NLOCAL];
+        	ModuleBase::GlobalFunc::ZEROS(ctot[i], GlobalV::NLOCAL);
+    	}
+    	ModuleBase::Memory::record("Pdiag_Double","ctot",GlobalV::NBANDS*GlobalV::NLOCAL,"cdouble");
+	}
+
+	k=0;
+    if (myid==0)
+    {
+        // mohan add nbnd0 2010-07-02
+        int nbnd0 = -1;
+        if (GlobalV::NBANDS < pv->loc_sizes[0])
+        {
+			// means all bands in this processor
+			// is needed ( is occupied)
+            nbnd0 = GlobalV::NBANDS;
+        }
+        else
+        {
+			// means this processor only save
+			// part of GlobalV::NBANDS.
+            nbnd0 = pv->loc_sizes[0];
+        }
+        if(pv->testpb)GlobalV::ofs_running << " nbnd in processor 0 is " << nbnd0 << std::endl;
+
+        for (i=0; i<nbnd0; i++)
+        {
+            for (j=0; j<GlobalV::NLOCAL; j++)
+            {
+				// change the order in processor 0.
+				// the contribution from processor 0.
+                ctot[k][j]=Z[j*pv->loc_sizes[0]+i];
+            }
+            k++;
+        }
+		// Z is useless in processor 0 now.
+		delete[] Z;
+    }
+    MPI_Barrier(comm);
+
+	for (i=1; i<= pv->lastband_in_proc; i++)
+    {
+        // mohan fix bug 2010-07-02
+        // rows indicates the data structure of Z.
+        // mpi_times indicates the data distribution
+        // time, each time send a band.
+        int rows = pv->loc_sizes[i];
+        int mpi_times;
+        if (i==pv->lastband_in_proc)
+        {
+            mpi_times = pv->lastband_number;
+        }
+        else
+        {
+            mpi_times = pv->loc_sizes[i];
+        }
+        if(pv->testpb)GlobalV::ofs_running << " nbnd in processor " << i << " is " << mpi_times << std::endl;
+        if (myid==i)
+        {
+            for (j=0; j<mpi_times; j++)
+            {
+                int tag = j;
+                std::complex<double> *send = new std::complex<double>[n];
+                int count = 0;
+
+                for (int m=0; m<rows*n; m++)
+                {
+                    if (m%rows==j)
+                    {
+                        send[count] = Z[m];
+                        ++count;
+                    }
+                }
+
+				// send the data to processor 0.
+                MPI_Send(send,n,mpicomplex,0,tag,comm);
+
+                delete[] send;
+            }
+			// third part to delete Z;
+			delete[] Z;
+        }
+        else if (myid==0)
+        {
+            int col=0;
+            for (j=0; j<mpi_times; j++)
+            {
+                std::complex<double> *ctmp = new std::complex<double>[GlobalV::NLOCAL];
+                ModuleBase::GlobalFunc::ZEROS(ctmp, GlobalV::NLOCAL);
+                int tag = j;
+                
+				// Processor 0 receive the data from other processors.
+				MPI_Recv(ctmp,n,mpicomplex,i,tag,comm,&status);
+
+                for (int m=0; m<GlobalV::NLOCAL; m++)
+                {
+                    ctot[k][m]=ctmp[m];
+//					GlobalV::ofs_running << " receive Z=" << ctmp[m] << std::endl;
+                }
+                k++;
+
+                delete[] ctmp;
+            }
+        }
+        //MPI_Barrier(comm);
+    }
+    if(pv->testpb)ModuleBase::GlobalFunc::OUT(GlobalV::ofs_running,"Final k",k);
+
+	// output the wave function if required.
+	// this is a bad position to output wave functions.
+	// but it works!
+	std::stringstream ss;
+	ss << GlobalV::global_out_dir << "LOWF_K_" << ik+1 << ".dat";
+    if(this->out_lowf)
+	{
+//		std::cout << " write the wave functions" << std::endl;
+		WF_Local::write_lowf_complex( ss.str(), ctot, ik );//mohan add 2010-09-09        
+	}
+
+	// mohan add 2010-09-10
+	// distribution of local wave functions 
+	// to each processor.
+	WF_Local::distri_lowf_complex( ctot, cc);
+	
+	// clean staff.
+	if(GlobalV::DRANK==0)
+	{
+    	for (int i=0; i<GlobalV::NBANDS; i++)
+    	{
+        	delete[] ctot[i];
+    	}
+    	delete[] ctot;
+	}
+
+    time_t time_end = time(NULL);
+//    GlobalV::ofs_running << " End   gath_eig_complex Time : " << ctime(&time_end);
+
+	ModuleBase::GlobalFunc::OUT_TIME("gather eigenvalues",time_start,time_end);
+    return;
+}
+#endif
+
+#ifdef __MPI
+void Pdiag_Double::gath_full_eig(MPI_Comm comm,int n,double **c,double *Z)
+{
+    ModuleBase::TITLE("Pdiag_Double","gath_full_eig");
+
+    time_t time_start = time(NULL);
+    //GlobalV::ofs_running << " Start gath_full_eig Time : " << ctime(&time_start);
+
+    int i,j,k,incx=1;
+    int *loc_sizes,loc_size,nprocs,myid;
+    MPI_Status status;
+    MPI_Comm_size(comm,&nprocs);
+    MPI_Comm_rank(comm,&myid);
+    loc_sizes =(int*)malloc(sizeof(int)*nprocs);
+    loc_size=n/nprocs;
+    for (i=0; i<nprocs; i++)
+    {
+        if (i<n%nprocs)
+        {
+            loc_sizes[i]=loc_size+1;
+        }
+        else
+        {
+            loc_sizes[i]=loc_size;
+        }
+    }
+    if (myid==0)
+    {
+        k=0;
+        for (i=0; i<loc_sizes[0]; i++)
+        {
+            for (j=0; j<n; j++)
+            {
+                // i : column index;
+                // j : row index.
+//              c[k*n+j]=Z[i*n+j];
+                c[k][j]=Z[j*loc_sizes[0]+i];
+                //			GlobalV::ofs_running << " Z=" << Z[i*n+j] << std::endl;
+            }
+            k++;
+        }
+    }
+    MPI_Barrier(comm);
+
+    for (i=1; i<nprocs; i++)
+    {
+        const int rows = loc_sizes[i];
+        const int mpi_times = rows;
+        if (myid==i)
+        {
+            for (j=0; j<mpi_times; j++)
+            {
+                int tag = j;
+
+                double *send = new double[n];
+
+                int count = 0;
+                for (int m=0; m<rows*n; m++)
+                {
+                    if (m%rows==j)
+                    {
+                        send[count] = Z[m];
+                        ++count;
+                    }
+                }
+
+                MPI_Send(send,n,MPI_DOUBLE,0,tag,comm);
+
+                delete[] send;
+            }
+        }
+        else if (myid==0)
+        {
+            int col=0;
+            for (j=0; j<mpi_times; j++)
+            {
+                double *ctmp = new double[n];
+                ModuleBase::GlobalFunc::ZEROS(ctmp, n);
+                int tag = j;
+                MPI_Recv(ctmp,n,MPI_DOUBLE,i,tag,comm,&status);
+
+                for (int m=0; m<n; m++)
+                {
+                    c[k][m]=ctmp[m];
+//					GlobalV::ofs_running << " receive Z=" << ctmp[m] << std::endl;
+                }
+                k++;
+
+                delete[] ctmp;
+            }
+        }
+        MPI_Barrier(comm);
+    }
+
+    for (int i=0; i<GlobalV::NLOCAL; i++)
+    {
+        Parallel_Common::bcast_double(c[i],GlobalV::NLOCAL);
+    }
+
+    time_t time_end = time(NULL);
+//    GlobalV::ofs_running << " End   gath_full_eig Time : " << ctime(&time_end);
+
+    return;
+}
+
+void Pdiag_Double::gath_full_eig_complex(MPI_Comm comm,int n,std::complex<double> **c,std::complex<double> *Z)
+{
+    ModuleBase::TITLE("Pdiag_Double","gath_full_eig_complex");
+
+    time_t time_start = time(NULL);
+    //GlobalV::ofs_running << " Start gath_full_eig_complex Time : " << ctime(&time_start);
+
+    int i,j,k,incx=1;
+    int *loc_sizes,loc_size,nprocs,myid;
+    MPI_Status status;
+    MPI_Comm_size(comm,&nprocs);
+    MPI_Comm_rank(comm,&myid);
+    loc_sizes =(int*)malloc(sizeof(int)*nprocs);
+    loc_size=n/nprocs;
+    for (i=0; i<nprocs; i++)
+    {
+        if (i<n%nprocs)
+        {
+            loc_sizes[i]=loc_size+1;
+        }
+        else
+        {
+            loc_sizes[i]=loc_size;
+        }
+    }
+    if (myid==0)
+    {
+        k=0;
+        for (i=0; i<loc_sizes[0]; i++)
+        {
+            for (j=0; j<n; j++)
+            {
+                // i : column index;
+                // j : row index.
+//              c[k*n+j]=Z[i*n+j];
+                c[k][j]=Z[j*loc_sizes[0]+i];
+                //			GlobalV::ofs_running << " Z=" << Z[i*n+j] << std::endl;
+            }
+            k++;
+        }
+    }
+    MPI_Barrier(comm);
+
+	for (i=1; i<nprocs; i++)
+	{
+		const int rows = loc_sizes[i];
+		const int mpi_times = rows;
+		if (myid==i)
+		{
+			for (j=0; j<mpi_times; j++)
+			{
+				int tag = j;
+
+				std::complex<double> *send = new std::complex<double>[n];
+
+				int count = 0;
+				for (int m=0; m<rows*n; m++)
+				{
+					if (m%rows==j)
+					{
+						send[count] = Z[m];
+						++count;
+					}
+				}
+
+				MPI_Send(send,n,mpicomplex,0,tag,comm);
+
+				delete[] send;
+			}
+		}
+		else if (myid==0)
+		{
+			int col=0;
+			for (j=0; j<mpi_times; j++)
+			{
+				std::complex<double> *ctmp = new std::complex<double>[n];
+				ModuleBase::GlobalFunc::ZEROS(ctmp, n);
+				int tag = j;
+				MPI_Recv(ctmp,n,mpicomplex,i,tag,comm,&status);
+
+				for (int m=0; m<n; m++)
+				{
+					c[k][m]=ctmp[m];
+					//					GlobalV::ofs_running << " receive Z=" << ctmp[m] << std::endl;
+				}
+				k++;
+
+				delete[] ctmp;
+			}
+		}
+		MPI_Barrier(comm);
+	}
+
+	for (int i=0; i<GlobalV::NLOCAL; i++)
+	{
+		Parallel_Common::bcast_complex_double(c[i],GlobalV::NLOCAL);
+	}
+
+    time_t time_end = time(NULL);
+    //GlobalV::ofs_running << " End   gath_full_eig Time : " << ctime(&time_end);
+	ModuleBase::GlobalFunc::OUT_TIME("gather full eigenvalues",time_start,time_end);
+
+    return;
+}
+#endif
+
+//LiuXh add 2021-09-06, clear memory, totwfc and WFC_GAMMA_aug not used now
+#ifdef __MPI
+#include "../src_pw/occupy.h"
+void Pdiag_Double::gath_eig(MPI_Comm comm,int n,double *Z)
+{
+    ModuleBase::TITLE("Pdiag_Double","gath_eig");
+    time_t time_start = time(NULL);
+//  GlobalV::ofs_running << " Start gath_eig Time : " << ctime(&time_start);
+
+    const Parallel_Orbitals* pv = this->ParaV;
+    
+    int i, j, k;
+    int nprocs,myid;
+    MPI_Status status;
+    MPI_Comm_size(comm,&nprocs);
+    MPI_Comm_rank(comm,&myid);
+
+    double **ctot;
+
+	// mohan add 2010-07-03
+	// the occupied bands are useless
+	// for calculating charge density.
+	if(GlobalV::DRANK > pv->lastband_in_proc)
+	{
+		delete[] Z;
+	}
+
+	// first we need to collect all
+	// the occupied bands.
+	// GlobalV::NBANDS * GlobalV::NLOCAL	
+	if(myid==0)
+	{
+		ctot = new double*[GlobalV::NBANDS];
+    	for (int i=0; i<GlobalV::NBANDS; i++)
+    	{
+        	ctot[i] = new double[GlobalV::NLOCAL];
+        	ModuleBase::GlobalFunc::ZEROS(ctot[i], GlobalV::NLOCAL);
+    	}
+    	ModuleBase::Memory::record("Pdiag_Double","ctot",GlobalV::NBANDS*GlobalV::NLOCAL,"double");
+	}
+
+    k=0;
+    if (myid==0)
+    {
+        // mohan add nbnd0 2010-07-02
+        int nbnd0 = -1;
+        if (GlobalV::NBANDS < pv->loc_sizes[0])
+        {
+			// means all bands in this processor
+			// is needed ( is occupied)
+            nbnd0 = GlobalV::NBANDS;
+        }
+        else
+        {
+			// means this processor only save
+			// part of GlobalV::NBANDS.
+            nbnd0 = pv->loc_sizes[0];
+        }
+        if(pv->testpb) GlobalV::ofs_running << " nbnd in processor 0 is " << nbnd0 << std::endl;
+
+//printf("from 0 to %d\n",nbnd0-1);
+        for (i=0; i<nbnd0; i++)
+        {
+            for (j=0; j<GlobalV::NLOCAL; j++)
+            {
+				// change the order in processor 0.
+				// the contribution from processor 0.
+                ctot[k][j]=Z[j*pv->loc_sizes[0]+i];
+            }
+            k++;
+        }
+		// Z is useless in processor 0 now.
+		delete[] Z;
+    }
+    MPI_Barrier(comm);
+
+
+    for (i=1; i<= pv->lastband_in_proc; i++)
+    {
+        // mohan fix bug 2010-07-02
+        // rows indicates the data structure of Z.
+        // mpi_times indicates the data distribution
+        // time, each time send a band.
+        int rows = pv->loc_sizes[i];
+        int mpi_times;
+        if (i==pv->lastband_in_proc)
+        {
+            mpi_times = pv->lastband_number;
+        }
+        else
+        {
+            mpi_times = pv->loc_sizes[i];
+        }
+        if(pv->testpb)GlobalV::ofs_running << " nbnd in processor " << i << " is " << mpi_times << std::endl;
+        if (myid==i)
+        {
+            for (j=0; j<mpi_times; j++)
+            {
+                int tag = j;
+                double *send = new double[n];
+                int count = 0;
+
+                for (int m=0; m<rows*n; m++)
+                {
+                    if (m%rows==j)
+                    {
+                        send[count] = Z[m];
+                        ++count;
+                    }
+                }
+
+				// send the data to processor 0.
+                MPI_Send(send,n,MPI_DOUBLE,0,tag,comm);
+
+                delete[] send;
+            }
+			// third part to delete Z;
+			delete[] Z;
+        }
+        else if (myid==0)
+        {
+            int col=0;
+            for (j=0; j<mpi_times; j++)
+            {
+                double *ctmp = new double[GlobalV::NLOCAL];
+                ModuleBase::GlobalFunc::ZEROS(ctmp, GlobalV::NLOCAL);
+                int tag = j;
+                
+				// Processor 0 receive the data from other processors.
+				MPI_Recv(ctmp,n,MPI_DOUBLE,i,tag,comm,&status);
+
+                for (int m=0; m<GlobalV::NLOCAL; m++)
+                {
+                    ctot[k][m]=ctmp[m];
+//					GlobalV::ofs_running << " receive Z=" << ctmp[m] << std::endl;
+                }
+                k++;
+
+                delete[] ctmp;
+            }
+        }
+        //MPI_Barrier(comm);
+    }
+    if(pv->testpb)ModuleBase::GlobalFunc::OUT(GlobalV::ofs_running,"Final k",k);
+/*
+if(myid==0){
+	double *vect=new double[GlobalV::NLOCAL*GlobalV::NBANDS];
+	double *form=new double[GlobalV::NBANDS*GlobalV::NBANDS];
+	int x,y;
+	for(x=0;x<GlobalV::NBANDS;x++){
+		for(y=0;y<GlobalV::NLOCAL;y++){
+			vect[x*GlobalV::NLOCAL+y]=ctot[x][y];
+		}
+	}
+		char chT='T';
+		char chN='N';
+		int	ne = GlobalV::NBANDS;
+		int m1, n1, k1;
+		double ONE=1.0,ZERO=0.0;
+		m1 = ne;
+		n1 = ne;
+		k1 = GlobalV::NLOCAL;
+		dgemm_(&chT, &chN, &m1, &n1, &k1, &ONE, vect, &k1, vect, &k1, &ZERO, form, &m1);
+		double di=0.0,oth=0.0;
+		for(x=0;x<ne;x++){
+			for(y=0;y<ne;y++){
+				if(x==y){
+					di+=fabs(form[x*ne+y]);
+				}else{
+					oth+=fabs(form[x*ne+y]);
+				}
+			}
+		}
+		di-=ne;
+		printf("\n\n\ndi=%.16lf\n\n\nother=%.16lf\n\n\n",di,oth);
+		//assert(0>1);
+}
+*/
+MPI_Barrier(comm);
+
+	// mohan add 2010-09-10
+	// output the wave function if required.
+	// this is a bad position to output wave functions.
+	// but it works!
+    if(this->out_lowf)
+	{
+		// read is in ../src_algorithms/wf_local.cpp
+		std::stringstream ss;
+		ss << GlobalV::global_out_dir << "LOWF_GAMMA_S" << GlobalV::CURRENT_SPIN+1 << ".dat";
+		// mohan add 2012-04-03, because we need the occupations for the
+		// first iteration. 
+		Occupy::calculate_weights();
+		WF_Local::write_lowf( ss.str(), ctot );//mohan add 2010-09-09        
+	}
+
+	// mohan add 2010-09-10
+	// distribution of local wave functions 
+	// to each processor.
+	// only used for GlobalV::GAMMA_ONLY_LOCAL
+	//WF_Local::distri_lowf( ctot, wfc);
+
+
+	// clean staff.
+	if(myid==0)
+	{
+    	for (int i=0; i<GlobalV::NBANDS; i++)
+    	{
+        	delete[] ctot[i];
+    	}
+    	delete[] ctot;
+	}
+
+    time_t time_end = time(NULL);
+    //GlobalV::ofs_running << " End   gath_eig Time : " << ctime(&time_end);
+	ModuleBase::GlobalFunc::OUT_TIME("gather eigenvalues",time_start,time_end);
+    return;
+}
+#endif
